refactor CLI calls into repokid.commands module
