+++ repokid/filters/age/__init__.py
@@ -2,7 +2,7 @@ import datetime
 
 from dateutil.tz import tzlocal
 from repokid import LOGGER
-from repokid.cli.repokid_cli import Filter
+from repokid.filters import Filter
 
 
 class AgeFilter(Filter):
