+++ repokid/filters/optout/__init__.py
@@ -1,6 +1,6 @@
 import time
 
-from repokid.cli.repokid_cli import Filter
+from repokid.filters import Filter
 
 
 class OptOutFilter(Filter):
